llm-service/requirements.txt
httpx~=0.28.1
ollama~=0.4.8
python-dotenv~=1.1.0
fastapi~=0.115.12
pydantic~=2.11.4
supabase~=2.15.1